코드 간략화 및 폴더구조 변경
